+++ modelopt/torch/export/unified_export_hf_streaming.py
@@ -43,10 +43,10 @@ from .hf_export_prep import (
     _warn_on_unsynced_moe_gate_up,
     requantize_resmooth_fused_llm_layers,
 )
+from .hf_weight_export import _dispatch_export_handler
 from .quant_aware_conversion import build_reverse_name_mapper
 from .quant_utils import _postprocess_single_tensor, get_quant_config
 from .registry import ExportContext
-from .unified_export_hf import _dispatch_export_handler
 
 __all__ = ["_export_transformers_checkpoint_streaming"]
